+++ test/test_unit_github.py
@@ -4,7 +4,7 @@ import json
 
 from utils import get_signed_headers
 
-from constants import TARGET_URL, EMAIL, PREMIUM_EMAIL, ORG
+from constants import TARGET_URL, EMAIL, PREMIUM_EMAIL, ORG, BASIC_EMAIL
 
 
 class GitHubUnitTestSuite(unittest.TestCase):
@@ -36,8 +36,32 @@ class GitHubUnitTestSuite(unittest.TestCase):
     def test_retrieve_file_private_access_repo_path(self):
         print("Running test: Retrieve a private file from the team's project based on repo and path")
         signedHeaders = get_signed_headers(PREMIUM_EMAIL)
-        response = requests.get(f"{TARGET_URL}/api/user/{ORG}/connectors/github/file?repo=https://github.com/polyverse-appsec/sara/&path=README.md", headers=signedHeaders)
+        response = requests.get(f"{TARGET_URL}/api/user/{ORG}/connectors/github/file?uri=https://github.com/polyverse-appsec/sara/&path=README.md", headers=signedHeaders)
         self.assertEqual(response.status_code, 200)
+
+    def test_premium_user_access_to_private_repo_success(self):
+        print("Running test: Check that premiumuser  WITH access to private repo can see it")
+        signedHeaders = get_signed_headers(PREMIUM_EMAIL)
+        response = requests.get(f"{TARGET_URL}/api/user/{ORG}/connectors/github/access?uri=https://github.com/polyverse-appsec/sara", headers=signedHeaders)
+        self.assertEqual(response.status_code, 200)
+        response = response.json() if 'body' not in response else json.loads(response['body'])
+        self.assertTrue(response)
+
+    def test_premium_user_access_to_private_repo_fail(self):
+        print("Running test: Check that premium user WITHOUT access to private repo can see it - no app install")
+        signedHeaders = get_signed_headers(EMAIL)
+        response = requests.get(f"{TARGET_URL}/api/user/{ORG}/connectors/github/access?uri=https://github.com/polyverse-appsec/sara", headers=signedHeaders)
+        self.assertEqual(response.status_code, 500)
+        response = response.text if 'body' not in response else response['body']
+        self.assertTrue("GitHub App Installation not found" in response)
+
+    def test_basic_user_access_to_private_repo_fail(self):
+        print("Running test: Check that Basic user WITHOUT access to private repo can see it - no app install")
+        signedHeaders = get_signed_headers(BASIC_EMAIL)
+        response = requests.get(f"{TARGET_URL}/api/user/{ORG}/connectors/github/access?uri=https://github.com/polyverse-appsec/sara", headers=signedHeaders)
+        self.assertEqual(response.status_code, 500)
+        response = response.text if 'body' not in response else response['body']
+        self.assertTrue("GitHub App Installation not found" in response)
 
     def test_retrieve_folders(self):
         print("Running test: Retrieve all folders from a public project")